Add helper for hiding global bar
Currently the global bar is initially hidden, then shown using JavaScript - this
causes a shift in the layout of the page (aka jank) and means that the global
bar is not available for users without JavaScript.

To avoid this, a change will need to be made in Static to the make the global
bar be initially shown by default. The bar can then be hidden on pages that
don't need it - and doing this using CSS will avoid jank.

This change adds a helper that checks whether a page should hide the global bar;
the view can then add a class to the body element that will hide the global bar.

These changes won't have any effect until Static has been updated.